+++ src/app/components/item-list/components/item-add/item-add.component.html
@@ -1,1 +1,2 @@
-<p>item-add works!</p>
+<input [(ngModel)]="title" type="text" placeholder="タスクを入力してください" />
+<button (click)="addItem()">追加</button>
