+++ src/app/components/item-list/components/item-add/item-add.component.spec.ts
@@ -1,15 +1,28 @@
 import { async, ComponentFixture, TestBed } from "@angular/core/testing";
-
+import { FormsModule } from "@angular/forms";
+import { RouterTestingModule } from "@angular/router/testing";
+import { StoreModule } from "@ngrx/store";
+import { provideMockStore } from "@ngrx/store/testing";
+import { ItemListService } from "./../../item-list.service";
 import { ItemAddComponent } from "./item-add.component";
+
 
 describe("ItemAddComponent", () => {
   let component: ItemAddComponent;
   let fixture: ComponentFixture<ItemAddComponent>;
+  let itemListService: ItemListService;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ItemAddComponent]
+      declarations: [ItemAddComponent],
+      providers: [
+        { provide: itemListService, useClass: ItemListService },
+        provideMockStore({})
+      ],
+      imports: [FormsModule, RouterTestingModule, StoreModule]
     }).compileComponents();
+
+    itemListService = TestBed.get(ItemListService);
   }));
 
   beforeEach(() => {
@@ -20,5 +33,21 @@ describe("ItemAddComponent", () => {
 
   it("should create", () => {
     expect(component).toBeTruthy();
+  });
+
+  describe("addItem()", () => {
+    it("itemListService の addedItem が呼ばれること", () => {
+      spyOn(itemListService, "addedItem");
+      component.addItem();
+      expect(itemListService.addedItem).toHaveBeenCalled();
+    });
+
+    it("component.title が '' になること", () => {
+      spyOn(itemListService, "addedItem");
+      component.addItem();
+      fixture.detectChanges();
+
+      expect(component.title).toBeFalsy();
+    });
   });
 });
