+++ src/app/components/item-list/components/item-add/item-add.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from "@angular/core";
+import { ItemListService } from "../../item-list.service";
+import { getUniqueStr } from "../../../../shared/domains/unique_id_maker";
+import { Item } from "../../../../shared/models";
 
 @Component({
   selector: "app-item-add",
@@ -6,7 +9,19 @@ import { Component, OnInit } from "@angular/core";
   styleUrls: ["./item-add.component.sass"]
 })
 export class ItemAddComponent implements OnInit {
-  constructor() {}
+  title: string;
+
+  constructor(private itemListService: ItemListService) {}
 
   ngOnInit() {}
+
+  async addItem() {
+    const id = getUniqueStr();
+    const title = this.title;
+    const createdAt = new Date();
+    const newItem = new Item(id, title, createdAt);
+    await this.itemListService.addedItem(newItem);
+
+    this.title = "";
+  }
 }
