+++ src/app/components/item-list/item-list.component.html
@@ -1,3 +1,2 @@
-<input [(ngModel)]="title" type="text" placeholder="タスクを入力してください" />
-<button (click)="addItem()">追加</button>
+<app-item-add></app-item-add>
 <app-item></app-item>
