+++ src/app/components/item-list/item-list.component.ts
@@ -1,7 +1,4 @@
-import { Component, OnInit } from "@angular/core";
-import { Item } from "../../shared/models";
-import { ItemListService } from "./item-list.service";
-import { getUniqueStr } from "../../shared/domains/unique_id_maker";
+import { Component } from "@angular/core";
 
 @Component({
   selector: "app-item-list",
@@ -9,17 +6,5 @@ import { getUniqueStr } from "../../shared/domains/unique_id_maker";
   styleUrls: ["./item-list.component.sass"]
 })
 export class ItemListComponent {
-  title: string;
-
-  constructor(private itemListService: ItemListService) {}
-
-  async addItem() {
-    const id = getUniqueStr();
-    const title = this.title;
-    const createdAt = new Date();
-    const newItem = new Item(id, title, createdAt);
-    await this.itemListService.addedItem(newItem);
-
-    this.title = "";
-  }
+  constructor() {}
 }
